refactor: utils methods moved into the file
