+++ components/Downloads/Release/BitnessDropdown.tsx
@@ -7,37 +7,7 @@ import semVer from 'semver';
 import Select from '@/components/Common/Select';
 import { useDetectOS } from '@/hooks/react-client';
 import { useReleaseContext } from '@/providers/releaseProvider';
-import type { UserOS } from '@/types/userOS';
-
-const installerBitnessMap = (os: UserOS, hasWindowsArm64: boolean) => {
-  const bitnessMap = {
-    WIN: [
-      {
-        label: '32-bit',
-        value: '86',
-      },
-      {
-        label: '64-bit',
-        value: '64',
-      },
-      {
-        label: 'ARM64',
-        value: 'arm64',
-        disabled: !hasWindowsArm64,
-      },
-    ],
-    MAC: [
-      {
-        label: '64-bit / ARM64',
-        value: '64',
-      },
-    ],
-    LINUX: [],
-    OTHER: [],
-  };
-
-  return bitnessMap[os];
-};
+import { installerBitnessMap } from '@/util/downloadUtils';
 
 const BitnessDropdown: FC = () => {
   const { bitness: userBitness } = useDetectOS();
