+++ components/withDownloadCategories.tsx
@@ -4,24 +4,9 @@ import type { FC, PropsWithChildren } from 'react';
 import { useClientContext } from '@/hooks/react-server';
 import getReleaseData from '@/next-data/releaseData';
 import ReleaseProvider from '@/providers/releaseProvider';
+import { getDownloadCategory, mapCategoriesToTabs } from '@/util/downloadUtils';
 
 import LinkTabs from './Common/LinkTabs';
-
-const getDownloadCategory = (pathname: string) => {
-  const segments = pathname
-    .replace('/new-design', '')
-    .split('/')
-    .filter(Boolean);
-  const [, c] = segments;
-
-  if (c === 'current' || typeof c === 'undefined') {
-    segments.unshift('download');
-  }
-
-  const [page, category, subCategory] = segments;
-
-  return { page, category, subCategory };
-};
 
 const WithDownloadCategories: FC<PropsWithChildren> = async ({ children }) => {
   const t = await getTranslations();
@@ -29,25 +14,28 @@ const WithDownloadCategories: FC<PropsWithChildren> = async ({ children }) => {
   const { pathname } = useClientContext();
   const { page, category, subCategory } = getDownloadCategory(pathname);
 
-  const mapCategoriesToTabs = (categories: Array<Array<string>>) =>
-    categories.map(([page, category, subCategory]) => ({
-      key: category,
-      label: t(`layouts.download.categories.${category}`),
-      link:
-        category === 'download'
-          ? `/${[page, subCategory].join('/')}`
-          : `/${[page, category, subCategory].join('/')}`,
-    }));
-
   return (
     <ReleaseProvider releases={releases}>
       <LinkTabs
         label={t('layouts.download.selectCategory')}
-        tabs={mapCategoriesToTabs([
-          [page, 'download', subCategory],
-          [page, 'package-manager', subCategory],
-          [page, 'source-code', subCategory],
-        ])}
+        tabs={mapCategoriesToTabs({
+          page: page,
+          categories: [
+            {
+              category: 'download',
+              label: t('layouts.download.categories.download'),
+            },
+            {
+              category: 'package-manager',
+              label: t('layouts.download.categories.package-manager'),
+            },
+            {
+              category: 'source-code',
+              label: t('layouts.download.categories.source-code'),
+            },
+          ],
+          subCategory: subCategory,
+        })}
         activeTab={category}
       >
         {children}
